Move all local to filesystems
