+++ src/extension.ts
@@ -11,8 +11,8 @@ import { GlobalStorage } from "./api/Storage";
 import { Tools } from "./api/Tools";
 import * as Debug from './api/debug';
 import { parseErrors } from "./api/errors/parser";
-import { DeployTools } from "./filesystems/deployTools";
-import { Deployment } from "./filesystems/deployment";
+import { DeployTools } from "./filesystems/local/deployTools";
+import { Deployment } from "./filesystems/local/deployment";
 import { CopyToImport } from "./components/copyToImport";
 import { CustomQSh } from "./components/cqsh";
 import { GetMemberInfo } from "./components/getMemberInfo";
