+++ src/filesystems/local/deployTools.ts
@@ -1,11 +1,11 @@
 import createIgnore, { Ignore } from 'ignore';
 import path, { basename } from 'path';
 import vscode, { Uri, WorkspaceFolder } from 'vscode';
-import { instance } from '../instantiate';
-import { DeploymentMethod, DeploymentParameters } from '../typings';
-import { LocalLanguageActions } from '../api/local/LocalLanguageActions';
+import { instance } from '../../instantiate';
+import { DeploymentMethod, DeploymentParameters } from '../../typings';
+import { LocalLanguageActions } from './LocalLanguageActions';
 import { Deployment } from './deployment';
-import { getGitAPI, md5Hash } from '../views/tools';
+import { getGitAPI, md5Hash } from '../../views/tools';
 
 type ServerFileChanges = { uploads: Uri[], relativeRemoteDeletes: string[] };
