+++ src/testing/action.ts
@@ -4,9 +4,9 @@ import vscode from "vscode";
 import { TestSuite } from ".";
 import { CompileTools } from "../api/CompileTools";
 import { Tools } from "../api/Tools";
-import { LocalLanguageActions } from "../api/local/LocalLanguageActions";
-import { DeployTools } from "../filesystems/deployTools";
-import { getEnvConfig } from "../api/local/env";
+import { LocalLanguageActions } from "../filesystems/local/LocalLanguageActions";
+import { DeployTools } from "../filesystems/local/deployTools";
+import { getEnvConfig } from "../filesystems/local/env";
 import { getMemberUri, getUriFromPath } from "../filesystems/qsys/QSysFs";
 import { instance } from "../instantiate";
 import { Action, IBMiObject } from "../typings";
