+++ src/views/diagnostics.ts
@@ -3,7 +3,7 @@ import * as vscode from "vscode";
 import { FileError } from "../typings";
 import { GlobalConfiguration } from "../api/Configuration";
 import Instance from "../Instance";
-import { getEvfeventFiles } from "../api/local/actions";
+import { getEvfeventFiles } from "../filesystems/local/actions";
 import { parseErrors } from "../api/errors/parser";
 import { findExistingDocumentByName, findExistingDocumentUri } from "./tools";
